backend/requirements.txt
Django==4.2.13
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
mysqlclient==2.2.4
Pillow==10.3.0
python-decouple==3.8
django-filter==23.5
